# swagger.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kasir Laundry Pro",
    "version": "1.0.0",
    "description": "API Routes and schema details of Kasir Laundry Pro Services"
  },
  "servers": [
    {
      "url": "https://kasirlaundrypro.netlify.app/api"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Logs in a user",
        "description": "Logs in a user with credentials",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User's email address"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Login berhasil."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email tidak ditemukan."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logs out a user",
        "description": "Logs out a user and invalidates their session or token",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful logout",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logout berhasil."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Terjadi kesalahan pada server."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "Signs up a new user",
        "description": "Creates a new user account",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User's email address"
                  },
                  "phone_number": {
                    "type": "string",
                    "description": "User's phone number"
                  },
                  "name": {
                    "type": "string",
                    "description": "User's name"
                  },
                  "password": {
                    "type": "string",
                    "description": "User's password"
                  }
                },
                "required": [
                  "email",
                  "phone_number",
                  "name",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Signup berhasil."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Email sudah digunakan."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "summary": "Changes the user's password",
        "description": "Allows a user to change their password",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "User's email address"
                  },
                  "currentPassword": {
                    "type": "string",
                    "description": "User's current password"
                  },
                  "newPassword": {
                    "type": "string",
                    "description": "User's new password"
                  }
                },
                "required": [
                  "email",
                  "currentPassword",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password successfully changed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Password updated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request, invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Current password is incorrect."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Current password is incorrect."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/google/callback": {
      "get": {
        "summary": "Handles Google OAuth callback",
        "description": "Handles the callback from Google after authentication",
        "tags": [
          "Auth"
        ],
        "responses": {
          "302": {
            "description": "Redirects to the frontend application with an authentication token"
          },
          "500": {
            "description": "Terjadi kesalahan server"
          }
        }
      }
    },
    "/auth/verify-email": {
      "get": {
        "summary": "Verify user email",
        "tags": [
          "Auth"
        ],
        "description": "Validates the email verification token and updates the user's status to 'verified'.",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The email verification token"
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully."
          },
          "400": {
            "description": "Invalid or expired verification link."
          },
          "500": {
            "description": "Terjadi kesalahan server."
          }
        }
      }
    },
...
  },
...
}